:arrow_up: moved mx records to delivery lib + added tlsa lookup.
